add blockscout, refactor CLI
