--- book/src/framework.md
@@ -1,1 +1,0 @@
-# Framework
