+++ framework/cmd/.gitignore
@@ -1,0 +1,4 @@
+blockscout/services/blockscout-db-data
+blockscout/services/logs
+blockscout/services/redis-data
+blockscout/services/stats-db-data
