+++ framework/cmd/main.go
@@ -4,14 +4,11 @@ import (
 	"embed"
 	"fmt"
 	"github.com/pelletier/go-toml"
-	"github.com/rs/zerolog"
 	"github.com/smartcontractkit/chainlink-testing-framework/framework"
 	"github.com/urfave/cli/v2"
 	"io/fs"
-	"io/ioutil"
 	"log"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strings"
 )
@@ -73,14 +70,49 @@ func main() {
 					{
 						Name:        "up",
 						Usage:       "ctf obs up",
+						Aliases:     []string{"u"},
 						Description: "Spins up a local observability stack: Grafana, Loki, Pyroscope",
 						Action:      func(c *cli.Context) error { return observabilityUp() },
 					},
 					{
 						Name:        "down",
 						Usage:       "ctf obs down",
+						Aliases:     []string{"d"},
 						Description: "Removes local observability stack",
 						Action:      func(c *cli.Context) error { return observabilityDown() },
+					},
+				},
+			},
+			{
+				Name:    "blockscout",
+				Aliases: []string{"bs"},
+				Usage:   "Controls local Blockscout stack",
+				Subcommands: []*cli.Command{
+					{
+						Name:        "up",
+						Usage:       "ctf bs up",
+						Aliases:     []string{"u"},
+						Description: "Spins up Blockscout stack",
+						Action:      func(c *cli.Context) error { return blockscoutUp() },
+					},
+					{
+						Name:        "down",
+						Usage:       "ctf bs down",
+						Aliases:     []string{"d"},
+						Description: "Removes Blockscout stack, wipes all Blockscout databases data",
+						Action:      func(c *cli.Context) error { return blockscoutDown() },
+					},
+					{
+						Name:        "reboot",
+						Usage:       "ctf bs reboot or ctf bs r",
+						Aliases:     []string{"r"},
+						Description: "Reboots Blockscout stack",
+						Action: func(c *cli.Context) error {
+							if err := blockscoutDown(); err != nil {
+								return err
+							}
+							return blockscoutUp()
+						},
 					},
 				},
 			},
@@ -91,25 +123,6 @@ func main() {
 	if err != nil {
 		log.Fatal(err)
 	}
-}
-
-func cleanDockerResources() error {
-	framework.L.Info().Str("label", "framework=ctf").Msg("Cleaning up docker containers")
-	// Bash command for removing Docker containers and networks with "framework=ctf" label
-	cmd := exec.Command("bash", "-c", `
-		docker ps -aq --filter "label=framework=ctf" | xargs -r docker rm -f && \
-		docker network ls --filter "label=framework=ctf" -q | xargs -r docker network rm
-	`)
-	framework.L.Debug().Msg("Running command")
-	if framework.L.GetLevel() == zerolog.DebugLevel {
-		fmt.Println(cmd.String())
-	}
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("error running clean command: %s", string(output))
-	}
-	framework.L.Info().Msgf("Done")
-	return nil
 }
 
 // extractAllFiles goes through the embedded directory and extracts all files to the current directory
@@ -155,7 +168,7 @@ func extractAllFiles(embeddedDir string) error {
 		}
 
 		// Write the file content to the target path
-		err = ioutil.WriteFile(targetPath, content, 0644)
+		err = os.WriteFile(targetPath, content, 0777)
 		if err != nil {
 			return fmt.Errorf("failed to write file %s: %w", targetPath, err)
 		}
@@ -163,50 +176,6 @@ func extractAllFiles(embeddedDir string) error {
 	})
 
 	return err
-}
-
-func observabilityUp() error {
-	framework.L.Info().Msg("Creating local observability stack")
-	if err := extractAllFiles("observability"); err != nil {
-		return err
-	}
-	cmd := exec.Command("bash", "-c", fmt.Sprintf(`
-		cd %s && \
-		docker compose up -d
-	`, "compose"))
-	framework.L.Debug().Msg("Running command")
-	if framework.L.GetLevel() == zerolog.DebugLevel {
-		fmt.Println(cmd.String())
-	}
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("error running clean command: %s", string(output))
-	}
-	framework.L.Info().Msg("Done")
-	fmt.Println()
-	framework.L.Info().Msgf("Loki: %s", LocalLogsURL)
-	framework.L.Info().Msgf("All logs: %s", "{job=\"ctf\"}")
-	framework.L.Info().Msgf("By log level: %s", "{job=\"ctf\", container=~\"node.*\"} |= \"WARN|INFO|DEBUG\"")
-	framework.L.Info().Msgf("Pyroscope: %s", LocalPyroScopeURL)
-	return nil
-}
-
-func observabilityDown() error {
-	framework.L.Info().Msg("Removing local observability stack")
-	cmd := exec.Command("bash", "-c", fmt.Sprintf(`
-		cd %s && \
-		docker compose down -v
-	`, "compose"))
-	framework.L.Debug().Msg("Running command")
-	if framework.L.GetLevel() == zerolog.DebugLevel {
-		fmt.Println(cmd.String())
-	}
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("error running clean command: %s", string(output))
-	}
-	framework.L.Info().Msg("Done")
-	return nil
 }
 
 // PrettyPrintTOML pretty prints TOML
@@ -220,7 +189,6 @@ func PrettyPrintTOML(inputFile string, outputFile string) error {
 		return fmt.Errorf("error parsing TOML: %v", err)
 	}
 
-	// Write the result to a new file
 	dumpData, err := tree.ToTomlString()
 	if err != nil {
 		return fmt.Errorf("error converting to TOML string: %v", err)
@@ -232,54 +200,4 @@ func PrettyPrintTOML(inputFile string, outputFile string) error {
 	}
 	framework.L.Info().Str("File", outputFile).Msg("File cleaned up and saved")
 	return nil
-}
-
-// BuildDocker runs Docker commands to set up a local registry, build an image, and push it.
-func BuildDocker(dockerfile string, buildContext string, imageName string) error {
-	registryRunning := isContainerRunning("local-registry")
-	if registryRunning {
-		fmt.Println("Local registry container is already running.")
-	} else {
-		framework.L.Info().Msg("Starting local registry container...")
-		err := runCommand("docker", "run", "-d", "-p", "5050:5000", "--name", "local-registry", "registry:2")
-		if err != nil {
-			return fmt.Errorf("failed to start local registry: %w", err)
-		}
-		framework.L.Info().Msg("Local registry started")
-	}
-
-	img := fmt.Sprintf("localhost:5050/%s:latest", imageName)
-	framework.L.Info().Str("DockerFile", dockerfile).Str("Context", buildContext).Msg("Building Docker image")
-	err := runCommand("docker", "build", "-t", fmt.Sprintf("localhost:5050/%s:latest", imageName), "-f", dockerfile, buildContext)
-	if err != nil {
-		return fmt.Errorf("failed to build Docker image: %w", err)
-	}
-	framework.L.Info().Msg("Docker image built successfully")
-
-	framework.L.Info().Str("Image", img).Msg("Pushing Docker image to local registry")
-	fmt.Println("Pushing Docker image to local registry...")
-	err = runCommand("docker", "push", img)
-	if err != nil {
-		return fmt.Errorf("failed to push Docker image: %w", err)
-	}
-	framework.L.Info().Msg("Docker image pushed successfully")
-	return nil
-}
-
-// isContainerRunning checks if a Docker container with the given name is running.
-func isContainerRunning(containerName string) bool {
-	cmd := exec.Command("docker", "ps", "--filter", fmt.Sprintf("name=%s", containerName), "--format", "{{.Names}}")
-	output, err := cmd.Output()
-	if err != nil {
-		return false
-	}
-	return strings.Contains(string(output), containerName)
-}
-
-// runCommand executes a command and prints the output.
-func runCommand(name string, args ...string) error {
-	cmd := exec.Command(name, args...)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	return cmd.Run()
 }
