+++ framework/cmd/observability/blockscout/envs/common-frontend.env
@@ -1,0 +1,17 @@
+NEXT_PUBLIC_API_HOST=localhost
+NEXT_PUBLIC_API_PROTOCOL=http
+NEXT_PUBLIC_STATS_API_HOST=http://localhost:8080
+NEXT_PUBLIC_NETWORK_NAME=Awesome chain
+NEXT_PUBLIC_NETWORK_SHORT_NAME=Awesome chain
+NEXT_PUBLIC_NETWORK_ID=5
+NEXT_PUBLIC_NETWORK_CURRENCY_NAME=Ether
+NEXT_PUBLIC_NETWORK_CURRENCY_SYMBOL=ETH
+NEXT_PUBLIC_NETWORK_CURRENCY_DECIMALS=18
+NEXT_PUBLIC_API_BASE_PATH=/
+NEXT_PUBLIC_APP_HOST=localhost
+NEXT_PUBLIC_APP_PROTOCOL=http
+NEXT_PUBLIC_HOMEPAGE_CHARTS=['daily_txs']
+NEXT_PUBLIC_VISUALIZE_API_HOST=http://localhost:8081
+NEXT_PUBLIC_IS_TESTNET=true
+NEXT_PUBLIC_API_WEBSOCKET_PROTOCOL=ws
+NEXT_PUBLIC_API_SPEC_URL=https://raw.githubusercontent.com/blockscout/blockscout-api-v2-swagger/main/swagger.yaml
