+++ framework/cmd/observability/blockscout/envs/common-smart-contract-verifier.env
@@ -1,0 +1,39 @@
+# Those are examples of existing configuration variables and their default values.
+# When uncommented, they would overwrite corresponding values from `base.toml`
+# configuration file.
+
+SMART_CONTRACT_VERIFIER__SERVER__HTTP__ENABLED=true
+SMART_CONTRACT_VERIFIER__SERVER__HTTP__ADDR=0.0.0.0:8050
+SMART_CONTRACT_VERIFIER__SERVER__HTTP__MAX_BODY_SIZE=2097152
+
+SMART_CONTRACT_VERIFIER__SERVER__GRPC__ENABLED=false
+SMART_CONTRACT_VERIFIER__SERVER__GRPC__ADDR=0.0.0.0:8051
+
+SMART_CONTRACT_VERIFIER__SOLIDITY__ENABLED=true
+SMART_CONTRACT_VERIFIER__SOLIDITY__COMPILERS_DIR=/tmp/solidity-compilers
+SMART_CONTRACT_VERIFIER__SOLIDITY__REFRESH_VERSIONS_SCHEDULE=0 0 * * * * *
+
+# It depends on the OS you are running the service on
+SMART_CONTRACT_VERIFIER__SOLIDITY__FETCHER__LIST__LIST_URL=https://solc-bin.ethereum.org/linux-amd64/list.json
+#SMART_CONTRACT_VERIFIER__SOLIDITY__FETCHER__LIST__LIST_URL=https://solc-bin.ethereum.org/macosx-amd64/list.json
+#SMART_CONTRACT_VERIFIER__SOLIDITY__FETCHER__LIST__LIST_URL=https://solc-bin.ethereum.org/windows-amd64/list.json
+
+SMART_CONTRACT_VERIFIER__VYPER__ENABLED=true
+SMART_CONTRACT_VERIFIER__VYPER__COMPILERS_DIR=/tmp/vyper-compilers
+SMART_CONTRACT_VERIFIER__VYPER__REFRESH_VERSIONS_SCHEDULE=0 0 * * * * *
+
+# It depends on the OS you are running the service on
+SMART_CONTRACT_VERIFIER__VYPER__FETCHER__LIST__LIST_URL=https://raw.githubusercontent.com/blockscout/solc-bin/main/vyper.list.json
+#SMART_CONTRACT_VERIFIER__VYPER__FETCHER__LIST__LIST_URL=https://raw.githubusercontent.com/blockscout/solc-bin/main/vyper.macos.list.json
+
+SMART_CONTRACT_VERIFIER__SOURCIFY__ENABLED=true
+SMART_CONTRACT_VERIFIER__SOURCIFY__API_URL=https://sourcify.dev/server/
+SMART_CONTRACT_VERIFIER__SOURCIFY__VERIFICATION_ATTEMPTS=3
+SMART_CONTRACT_VERIFIER__SOURCIFY__REQUEST_TIMEOUT=10
+
+SMART_CONTRACT_VERIFIER__METRICS__ENABLED=false
+SMART_CONTRACT_VERIFIER__METRICS__ADDR=0.0.0.0:6060
+SMART_CONTRACT_VERIFIER__METRICS__ROUTE=/metrics
+
+SMART_CONTRACT_VERIFIER__JAEGER__ENABLED=false
+SMART_CONTRACT_VERIFIER__JAEGER__AGENT_ENDPOINT=localhost:6831
