+++ framework/cmd/observability/blockscout/envs/common-stats.env
@@ -1,0 +1,27 @@
+# Those are examples of existing configuration variables and their default values.
+# When uncommented, they would overwrite corresponding values from `base.toml`
+# configuration file.
+
+STATS__SERVER__HTTP__ENABLED=true
+STATS__SERVER__HTTP__ADDR=0.0.0.0:8050
+STATS__SERVER__HTTP__MAX_BODY_SIZE=2097152
+
+STATS__SERVER__GRPC__ENABLED=false
+STATS__SERVER__GRPC__ADDR=0.0.0.0:8051
+
+STATS__DB_URL=
+STATS__BLOCKSCOUT_DB_URL=
+STATS__CREATE_DATABASE=false
+STATS__RUN_MIGRATIONS=false
+STATS__DEFAULT_SCHEDULE=0 0 1 * * * *
+STATS__FORCE_UPDATE_ON_START=false
+
+STATS__METRICS__ENABLED=false
+STATS__METRICS__ADDR=0.0.0.0:6060
+STATS__METRICS__ROUTE=/metrics
+
+STATS__JAEGER__ENABLED=false
+STATS__JAEGER__AGENT_ENDPOINT=localhost:6831
+
+STATS__TRACING__ENABLED=true
+STATS__TRACING__FORMAT=default
