+++ framework/cmd/observability/blockscout/envs/common-user-ops-indexer.env
@@ -1,0 +1,48 @@
+## Those are examples of existing configuration variables and their default values.
+## When uncommented, they would overwrite corresponding values from `base.toml`
+## configuration file.
+
+USER_OPS_INDEXER__SERVER__HTTP__ENABLED=true
+USER_OPS_INDEXER__SERVER__HTTP__ADDR=0.0.0.0:8050
+USER_OPS_INDEXER__SERVER__HTTP__MAX_BODY_SIZE=2097152
+USER_OPS_INDEXER__SERVER__GRPC__ENABLED=false
+USER_OPS_INDEXER__SERVER__GRPC__ADDR=0.0.0.0:8051
+
+USER_OPS_INDEXER__API__MAX_PAGE_SIZE=100
+
+## (required) no default value available
+USER_OPS_INDEXER__INDEXER__RPC_URL=""
+USER_OPS_INDEXER__INDEXER__CONCURRENCY=20
+USER_OPS_INDEXER__INDEXER__ENTRYPOINTS__V06=true
+USER_OPS_INDEXER__INDEXER__ENTRYPOINTS__V07=true
+
+USER_OPS_INDEXER__INDEXER__REALTIME__ENABLED=true
+
+USER_OPS_INDEXER__INDEXER__PAST_RPC_LOGS_INDEXER__ENABLED=true
+USER_OPS_INDEXER__INDEXER__PAST_RPC_LOGS_INDEXER__BLOCK_RANGE=1000
+
+USER_OPS_INDEXER__INDEXER__PAST_DB_LOGS_INDEXER__ENABLED=true
+USER_OPS_INDEXER__INDEXER__PAST_DB_LOGS_INDEXER__START_BLOCK=-100000
+USER_OPS_INDEXER__INDEXER__PAST_DB_LOGS_INDEXER__END_BLOCK=0
+
+## (required) no default value available
+USER_OPS_INDEXER__DATABASE__CONNECT__URL=""
+# OR
+#USER_OPS_INDEXER__DATABASE__CONNECT__KV__HOST=
+#USER_OPS_INDEXER__DATABASE__CONNECT__KV__PORT=
+#USER_OPS_INDEXER__DATABASE__CONNECT__KV__USER=
+#USER_OPS_INDEXER__DATABASE__CONNECT__KV__PASSWORD=
+#USER_OPS_INDEXER__DATABASE__CONNECT__KV__DBNAME=
+
+USER_OPS_INDEXER__DATABASE__CREATE_DATABASE=false
+USER_OPS_INDEXER__DATABASE__RUN_MIGRATIONS=false
+
+USER_OPS_INDEXER__METRICS__ENABLED=true
+USER_OPS_INDEXER__METRICS__ADDR=0.0.0.0:6060
+USER_OPS_INDEXER__METRICS__ROUTE=/metrics
+
+USER_OPS_INDEXER__JAEGER__ENABLED=false
+USER_OPS_INDEXER__JAEGER__AGENT_ENDPOINT=localhost:6831
+
+USER_OPS_INDEXER__TRACING__ENABLED=true
+USER_OPS_INDEXER__TRACING__FORMAT=default
