+++ framework/cmd/observability/blockscout/envs/common-visualizer.env
@@ -1,0 +1,1 @@
+VISUALIZER__SERVER__GRPC__ENABLED=false
