+++ framework/cmd/observability/blockscout/services/backend.yml
@@ -1,0 +1,16 @@
+version: '3.9'
+
+services:
+  backend:
+    image: blockscout/${DOCKER_REPO:-blockscout}:${DOCKER_TAG:-latest}
+    pull_policy: always
+    restart: always
+    stop_grace_period: 5m
+    container_name: 'backend'
+    command: sh -c "bin/blockscout eval \"Elixir.Explorer.ReleaseTasks.create_and_migrate()\" && bin/blockscout start"
+    extra_hosts:
+      - 'host.docker.internal:host-gateway'
+    env_file:
+      -  ../envs/common-blockscout.env
+    volumes:
+      - ./logs/:/app/logs/
