+++ framework/cmd/observability/blockscout/services/frontend.yml
@@ -1,0 +1,11 @@
+version: '3.9'
+
+services:
+  frontend:
+    image: ghcr.io/blockscout/frontend:${FRONTEND_DOCKER_TAG:-latest}
+    pull_policy: always
+    platform: linux/amd64
+    restart: always
+    container_name: 'frontend'
+    env_file:
+      -  ../envs/common-frontend.env
