+++ framework/cmd/observability/blockscout/services/sig-provider.yml
@@ -1,0 +1,9 @@
+version: '3.9'
+
+services:
+  sig-provider:
+    image: ghcr.io/blockscout/sig-provider:${SIG_PROVIDER_DOCKER_TAG:-latest}
+    pull_policy: always
+    platform: linux/amd64
+    restart: always
+    container_name: 'sig-provider'
