+++ framework/cmd/observability/blockscout/services/smart-contract-verifier.yml
@@ -1,0 +1,11 @@
+version: '3.9'
+
+services:
+  smart-contract-verifier:
+    image: ghcr.io/blockscout/smart-contract-verifier:${SMART_CONTRACT_VERIFIER_DOCKER_TAG:-latest}
+    pull_policy: always
+    platform: linux/amd64
+    restart: always
+    container_name: 'smart-contract-verifier'
+    env_file:
+      -  ../envs/common-smart-contract-verifier.env
