+++ framework/cmd/observability/blockscout/services/visualizer.yml
@@ -1,0 +1,11 @@
+version: '3.9'
+
+services:
+  visualizer:
+    image: ghcr.io/blockscout/visualizer:${VISUALIZER_DOCKER_TAG:-latest}
+    pull_policy: always
+    platform: linux/amd64
+    restart: always
+    container_name: 'visualizer'
+    env_file:
+      -  ../envs/common-visualizer.env
